requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
